test build for zip support

+++ packages/games/emulators/mupen64plussa/mupen64plussa-core/m64p.sh
@@ -36,8 +36,19 @@ if [[ ! -f "$M64PCONF" ]]; then
 	cp $SHARE/mupen64plus.cfg $M64PCONF
 fi
 
-rm -rf $TMP
+# Create a clean working directory
+if [ -d $TMP ]; then
+	rm -rf $TMP
+fi
 mkdir -p $TMP
+
+# Unzip or copy the rom to the working directory
+if [ `echo $2 | grep -i .zip | wc -l` -eq 1 ]; then
+	#unpack the zip file
+  	unzip -q -o "$2" -d $TMP
+else
+	cp $2 $TMP
+fi
 
 cp $M64PCONF $TMP
 SET_PARAMS="--set Core[SharedDataPath]=$TMP --set Video-Rice[ResolutionWidth]=$SCREENWIDTH"
@@ -94,19 +105,26 @@ else
 	SET_PARAMS="$SET_PARAMS --rsp mupen64plus-rsp-cxd4.so"
 fi
 
+# Get the true filename without extension
+ROM_W_EXT="${2##*/}"
+ROM="${FILE_W_EXT%%.*}"
+
 echo ${SET_PARAMS}
 
 case $1 in
 	"m64p_gliden64")
-		/usr/local/bin/mupen64plus --configdir $TMP --gfx mupen64plus-video-GLideN64 $SET_PARAMS "$2"
+		/usr/local/bin/mupen64plus --configdir $TMP --gfx mupen64plus-video-GLideN64 $SET_PARAMS "$TMP/$ROM"
 	;;
 	"m64p_gl64mk2")
-		/usr/local/bin/mupen64plus --configdir $TMP --gfx mupen64plus-video-glide64mk2 $SET_PARAMS "$2"
+		/usr/local/bin/mupen64plus --configdir $TMP --gfx mupen64plus-video-glide64mk2 $SET_PARAMS "$TMP/$ROM"
 	;;
 	"m64p_rice")
-		/usr/local/bin/mupen64plus --configdir $TMP --gfx mupen64plus-video-rice $SET_PARAMS "$2"
+		/usr/local/bin/mupen64plus --configdir $TMP --gfx mupen64plus-video-rice $SET_PARAMS "$TMP/$ROM"
 	;;
 	*)
-		/usr/local/bin/mupen64plus --configdir $TMP --gfx mupen64plus-video-rice $SET_PARAMS "$2"
+		/usr/local/bin/mupen64plus --configdir $TMP --gfx mupen64plus-video-rice $SET_PARAMS "$TMP/$ROM"
 	;;
 esac
+
+# Remove temporary dir
+mkdir -p $TMP
